♻️ refactor(developer): extract logic into custom hook and split UI components
